Flow for manipulation with password (via PasswordController)
