+++ src/main/java/edu/netcracker/project/logistic/config/security/SpringSecurityConfig.java
@@ -41,8 +41,8 @@ public class SpringSecurityConfig extends WebSecurityConfigurerAdapter {
     protected void configure(HttpSecurity http) throws Exception {
         http.csrf().disable()
                 .authorizeRequests()
-                .antMatchers("/", "/index", "/registration", "/registration/complete", "/registration/confirm", "/test", "/login/forgot/password", "/login/reset/password", "/login/reset/password/**").permitAll()
-                .antMatchers("/person_main").hasAnyRole("ADMIN", "MANAGER", "COURIER", "CALL_CENTER_AGENT", "USER")
+                .antMatchers("/", "/index", "/registration", "/registration/complete", "/registration/confirm", "/test", "/password/forgot", "/password/reset", "/password/reset/**").permitAll()
+                .antMatchers("/person_main", "/password/change").hasAnyRole("ADMIN", "MANAGER", "COURIER", "CALL_CENTER_AGENT", "USER")
                 .antMatchers("/admin/**").hasAnyRole("ADMIN")
                 .antMatchers("/manager/**").hasAnyRole("MANAGER")
                 .antMatchers("/courier/**").hasAnyRole("COURIER")
