+++ src/main/resources/templates/forgot_password.html
@@ -21,7 +21,7 @@
                 <p>When you fill in your registered email address, you will be sent instructions on how to reset your password.</p>
 
                 <form th:method="post"
-                      th:action="@{'/login/forgot/password'}"
+                      th:action="@{'/password/forgot'}"
                       th:object="${forgotPassword}">
                     <div class="form-group">
                         <input type="email" class="form-control" id="email" placeholder="email address" required="required"
