+++ src/main/resources/templates/fragments/left_menu.html
@@ -84,7 +84,7 @@
                                     <div class="panel-body">
                                         <ul class="nav navbar-nav">
                                             <li><a id="user_personal_area" th:href = "@{/user/personal}" href="#">Personal Area</a></li>
-                                            <li><a id="user_change_password" th:href="@{/user/change/password}" href="#">Change password</a></li>
+                                            <li><a id="user_change_password" th:href="@{/password/change}" href="#">Change password</a></li>
                                             <li><a id="user_my_orders" href="#">My orders</a></li>
                                             <li><a id="user_create_new_order" href="#">Create new order</a></li>
                                         </ul>
