+++ src/main/resources/templates/login.html
@@ -34,11 +34,7 @@
                             Your profile has been deleted successfully.
                         </div>
                     </div>
-                    <div th:if="${param.changeUsername}">
-                        <div class="alert alert-info">
-                            Your username has been successfully changed. Try log in with this new one.
-                        </div>
-                    </div>
+
 
                     <div class="form-group">
                         <label class="control-label col-sm-2" for="username">Login:</label>
@@ -76,7 +72,7 @@
             </form>
 
             <div class="etc-login-form">
-                <p>forgot your password? <a th:href="@{__${#httpServletRequest.requestURI}__/forgot/password}">click
+                <p>forgot your password? <a th:href="@{/password/forgot}">click
                     here</a></p>
                 <p>new user? <a th:href="@{/registration}">create new account</a></p>
             </div>
