+++ src/main/resources/templates/reset_password.html
@@ -14,7 +14,7 @@
 
             <h1 style="text-align: center">Create new password</h1>
 
-            <form style="" th:object="${resetPassword}" th:action="@{'/login/reset/password/' + ${resetPassword.personId}}" th:method="POST">
+            <form style="" th:object="${resetPassword}" th:action="@{'/password/reset' + ${resetPassword.personId}}" th:method="POST">
 
                 <div class="form-group required">
                     <label for="newPassword" class="control-label">New password</label>
