+++ src/main/resources/templates/user/user_change_password.html
@@ -30,7 +30,7 @@
                         </div>
                     </div>
 
-                    <form style="" th:object="${changePassword}" th:action="@{'/user/change/password'}" th:method="POST">
+                    <form style="" th:object="${changePassword}" th:action="@{'/password/change'}" th:method="POST">
                         <div class="form-group required"
                              th:classappend="${#fields.hasErrors('oldPassword')}? 'has-error'">
                             <label for="oldPassword" class="control-label">Old password</label>
